+++ addOns/selenium/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - Update minimum ZAP version to 2.9.0.
+- Set Firefox browser.tabs.documentchannel pref to false to fix HUD issue
 
 ## [15.1.0] - 2020-01-17
 ### Added